Replace Conversion class by boost::lexical_cast Issue : #5519

+++ src/39_map/DrawablePhysicalStop.h
@@ -55,7 +55,7 @@ public:
 
 private:
 
-    const int _physicalStopId;
+	const long long _physicalStopId;
     const std::string _name;
 	geos::geom::Coordinate _point;
